Use an 'unsigned int' as the rand_r seed.
This then matches the stdlib.h declaration.

Looking at the numbers generated, having that be a 16-bit value doesn't
appear to affect the quality of the output significantly. It's certainly
good enough for typical use cases.

+++ src/rand_r.rs
@@ -1,7 +1,7 @@
 //! Rust implementation of C library function `rand_r`
 //!
 //! Licensed under the Blue Oak Model Licence 1.0.0
-use core::ffi::{c_int, c_uint, c_ulong};
+use core::ffi::{c_int, c_uint};
 
 #[cfg_attr(not(feature = "rand_r"), export_name = "tinyrlibc_RAND_MAX")]
 #[cfg_attr(feature = "rand_r", no_mangle)]
@@ -12,20 +12,20 @@ pub static RAND_MAX: c_int = c_int::MAX;
 /// Passing NULL (core::ptr::null()) gives undefined behaviour.
 #[cfg_attr(not(feature = "rand_r"), export_name = "tinyrlibc_rand_r")]
 #[cfg_attr(feature = "rand_r", no_mangle)]
-pub unsafe extern "C" fn rand_r(seedp: *mut c_ulong) -> c_int {
-	let mut result: c_ulong;
+pub unsafe extern "C" fn rand_r(seedp: *mut c_uint) -> c_int {
+	let mut result: c_int;
 
-	fn pump(input: c_ulong) -> c_ulong {
+	fn pump(input: u32) -> u32 {
 		// This algorithm is mentioned in the ISO C standard
 		input.wrapping_mul(1103515245).wrapping_add(12345)
 	}
 
-	fn select_top(state: c_ulong, bits: usize) -> c_ulong {
+	fn select_top(state: u32, bits: usize) -> c_int {
 		// ignore the lower 16 bits, as they are low quality
-		(state >> 16) & ((1 << bits) - 1)
+		((state >> 16) & ((1 << bits) - 1)) as c_int
 	}
 
-	let mut next = *seedp;
+	let mut next = *seedp as u32;
 	if c_int::MAX == 32767 || cfg!(feature = "rand_max_i16") {
 		// pull 15 bits in one go
 		next = pump(next);
@@ -39,7 +39,7 @@ pub unsafe extern "C" fn rand_r(seedp: *mut c_ulong) -> c_int {
 		next = pump(next);
 		result |= select_top(next, 10);
 	}
-	*seedp = next;
+	*seedp = next as c_uint;
 
 	result as c_int
 }
